# openapi-spec.yaml
openapi: 3.0.0
info:
  title: Self Description Creator API
  version: 0.1.0
servers:
  - url: https://sd-creator.gxfs.gx4fm.org
    description: Demonstration server - maintained by msg.
paths:
  /federated-catalogue/self-descriptions:
    post:
      summary: Creates a sample Self Description based on provided Claims and adds it to configured GXFS Federated Catalogue instance.
      description: Issuer and Private Key used for creating the Proof are configured on server side.
      requestBody:
        description: JSON-LD Claims to be placed inside the Verifiable Credential of the resulting Self Description.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              {
                "id": "https://gaia-x.eu/.well-known/service1.json",
                "type": "gx:ServiceOffering",
                "gx:providedBy": {
                  "id": "https://gaia-x.eu/.well-known/participant.json"
                },
                "gx:policy": "",
                "gx:termsAndConditions": {
                  "gx:URL": "http://termsandconds.com",
                  "gx:hash": "d8402a23de560f5ab34b22d1a142feb9e13b3143"
                },
                "gx:dataAccountExport": {
                  "gx:requestType": "API",
                  "gx:accessType": "digital",
                  "gx:formatType": "application/json"
                }
              }
      responses:
        '201': # status code
          description: In case the Self Description has been created and successfully added to the configured GXFS Federated Catalogue instance.
          content:
            application/json:
              schema:
                type: object
              example:
                {"status": "success"}
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object

  /self-description:
    post:
      summary: Creates a sample Self Description based on provided Claims.
      description: Issuer and Private Key used to create the Proof are configured on the server side.
      requestBody:
        description: JSON-LD Claims to be placed inside the Verifiable Credential of the resulting Self Description.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              {
                "id": "https://gaia-x.eu/.well-known/service1.json",
                "type": "gx:ServiceOffering",
                "gx:providedBy": {
                  "id": "https://gaia-x.eu/.well-known/participant.json"
                },
                "gx:policy": "",
                "gx:termsAndConditions": {
                  "gx:URL": "http://termsandconds.com",
                  "gx:hash": "d8402a23de560f5ab34b22d1a142feb9e13b3143"
                },
                "gx:dataAccountExport": {
                  "gx:requestType": "API",
                  "gx:accessType": "digital",
                  "gx:formatType": "application/json"
                }
              }
      responses:
        '200': # status code
          description: The created Self Description.
          content:
            application/json:
              schema:
                type: object
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object

  /sd-from-vp-without-proof:
    post:
      summary: Creates a Self Description (Verifiable Presentation having a proof) based on a provided Verifiable Presentation without a proof.
      description: Issuer and Private Key used to create the Proof are configured on the server side.
      requestBody:
        description: Verifiable Presentation in JSON-LD format having no proof.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              {
                "@context": [
                    "https://www.w3.org/2018/credentials/v1"
                ],
                "holder": "did:web:compliance.lab.gaia-x.eu",
                "id": "http://example.org/presentations/3731",
                "type": [
                    "VerifiablePresentation"
                ],
                "verifiableCredential": [
                    {
                        "@context": [
                            "https://www.w3.org/2018/credentials/v1",
                            "https://www.w3.org/2018/credentials/examples/v1"
                        ],
                        "credentialSubject": {
                            "@context": {
                                "example": "https://github.com/GAIA-X4PLC-AAD/ontology-management/tree/main/examples/example_ontology/",
                                "openlabel": "https://openlabel.asam.net/V1-0-0/ontologies/",
                                "sh": "http://www.w3.org/ns/shacl#",
                                "skos": "http://www.w3.org/2004/02/skos/core#",
                                "xsd": "http://www.w3.org/2001/XMLSchema#"
                            },
                            "@id": "fdfgdfgdf",
                            "@type": "example:ExampleServiceOffering",
                            "example:belongs_to": {
                                "@id": "sdfgdfgdfg"
                            },
                            "example:hasJunctionIntersection": {
                                "@id": "serssefr"
                            },
                            "example:property1": {
                                "@type": "xsd:string",
                                "@value": "sdfsdf"
                            },
                            "example:property2": {
                                "@type": "xsd:string",
                                "@value": "svrvgerswv"
                            }
                        },
                        "expirationDate": "2022-06-18T12:30:00Z",
                        "issuanceDate": "2022-01-01T12:30:00Z",
                        "issuer": "did:web:compliance.lab.gaia-x.eu",
                        "proof": {
                            "created": "2022-01-01T12:30:00Z",
                            "jws": "eyJiNjQiOmZhbHNlLCJjcml0IjpbImI2NCJdLCJhbGciOiJQUzI1NiJ9..LyowYUxUDtOXur-JsLe_VSEgA9h1ERcHQbuq4Aq3wjrCnb6WIBlm-A1N3P4hVLuxVfWiBK-aNooxynbRAIToQT5sqI3onf6GLjPd53cUOAUnLie7kE7Sd4Vio_Gy085oS9lSc_wIiK6m7ebEownz9UNJuyFJLpKDIkJngZ0RNcsoIl2Ntny9lcXt6B0nDuHbmjfJZZAwqWEaLZTAllI2Fd4vKpOt037PbKDWSF3uk11Zaw5-T48X6E-RqA-fH5VRhKCxK6bpybt7vUfj4tEG0vC5eCou_uolfoCy3YSODcqmK4N2kb6vTDKnRrvIVYJWGrzSTbT9sjpbVeg3eQG3999iGgLWNyYj9k40EuSRfph7i97ODcC-CdkydQFZ_J2HBV6erS3oIPQk_To37Itx2QojK_GgUfiYYG8UQM5suBDKXaog9rYaOl0vP4W7HYqgDQjVVhN4uhhRSmtAD3ICy_Npfy4tccjAWw9_cPZNWY3wdIZpGQPo6yYjLPtRJx5rN1T8ECAQMRT-bnd33M06onzAUu1zDc6v_lyy2dxhtiAQQZH0d_mGVN_uH4_9yL-Tw7WO2B4u-TTICnU1JQ-RJ3gLC0PR9W1ph2EfJf5jYYOkJYYCf-1XQiktuJ9bcpn-OoSqdR5ilh0SWclqA3eu5RjBbI8FRlrgOIVR1UuVHfw",
                            "proofPurpose": "assertionMethod",
                            "type": "JsonWebSignature2020",
                            "verificationMethod": "did:web:compliance.lab.gaia-x.eu"
                        },
                        "type": [
                            "VerifiableCredential"
                        ]
                    }
                ]
            }
      responses:
        '200': # status code
          description: The created Self Description.
          content:
            application/json:
              schema:
                type: object
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object

  /vc-from-claims:
    post:
      summary: Creates a Verifiable Credential based on provided Claims.
      description: Issuer and Private Key used to create the Proof are configured on the server side.
      requestBody:
        description: JSON-LD Claims to be placed inside the Verifiable Credential.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              {
                "id": "https://gaia-x.eu/.well-known/service1.json",
                "type": "gx:ServiceOffering",
                "gx:providedBy": {
                  "id": "https://gaia-x.eu/.well-known/participant.json"
                },
                "gx:policy": "",
                "gx:termsAndConditions": {
                  "gx:URL": "http://termsandconds.com",
                  "gx:hash": "d8402a23de560f5ab34b22d1a142feb9e13b3143"
                },
                "gx:dataAccountExport": {
                  "gx:requestType": "API",
                  "gx:accessType": "digital",
                  "gx:formatType": "application/json"
                }
              }
      responses:
        '200': # status code
          description: The created Self Description.
          content:
            application/json:
              schema:
                type: object
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object

  /sd-from-vcs:
    post:
      summary: Creates a Verifiable Presentation/Self Description based on a provided list of Verifiable Credentials.
      description: Issuer and Private Key used to create the Proof are configured on the server side.
      requestBody:
        description: List of Verifiable Credentials to be placed inside the Verifiable Presentation/Self Description.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              [
                  {
                  "@context": [
                      "https://www.w3.org/2018/credentials/v1",
                      "https://www.w3.org/2018/credentials/examples/v1"
                  ],
                  "credentialSubject": {
                      "@context": {
                          "example": "https://github.com/GAIA-X4PLC-AAD/ontology-management/tree/main/examples/example_ontology/",
                          "openlabel": "https://openlabel.asam.net/V1-0-0/ontologies/",
                          "sh": "http://www.w3.org/ns/shacl#",
                          "skos": "http://www.w3.org/2004/02/skos/core#",
                          "xsd": "http://www.w3.org/2001/XMLSchema#"
                      },
                      "@id": "ololololoolollo",
                      "@type": "example:ExampleServiceOffering",
                      "example:belongs_to": {
                          "@id": "fghdhdfgrttz"
                      },
                      "example:hasJunctionIntersection": {
                          "@id": "fxghfhgfh"
                      },
                      "example:property1": {
                          "@type": "xsd:string",
                          "@value": "lololololol"
                      },
                      "example:property2": {
                          "@type": "xsd:string",
                          "@value": "lololloolol"
                      }
                  },
                  "expirationDate": "2022-06-18T12:30:00Z",
                  "issuanceDate": "2022-01-01T12:30:00Z",
                  "issuer": "did:web:compliance.lab.gaia-x.eu",
                  "proof": {
                      "created": "2022-01-01T12:30:00Z",
                      "jws": "eyJiNjQiOmZhbHNlLCJjcml0IjpbImI2NCJdLCJhbGciOiJQUzI1NiJ9..gv_mzzc9aoZbbVUpVZ3VDs3AAoLm9n9On5s9YNwBj-XqJx3gl2lJozy1uz22GhaaLJ3J1SQa_-diZFgIfi6TmCLnV78j2Xs61aNknCFhBWrvt_GrqV13dCIhRas1pxnBVLknBrhh-9bFj_b5_zTFGACNfBUsf2n4LnS-gk14JeyDc_T9GGTa1_tDqYljZZLonY9oe6raQ6NUdgNnAOX5KeHyWN7F0NGz4sXfiIsJU_GidjvTa4pyM7mQfpXpgX7paOENSkrWA3dDtlyxsyUmIRyAzGvizifVesQfaYucWKmpgvO2wMi_KmsRQmy2E2Xvb8dVkQ3hPl2BUiJVnNOcC3FyKkrFue-YIvZEKSjuYlX2c-EaK0sg5M0OkE7kd2VBT8Hy_UnM1qqMmHQQj2TaSnkr2T04B4qVLcHQVQ9qDP9JH-8g3BsGdFMZK-Xi_rKp1jk_lItCkd_dBTfFFIbGxMp-ckVr3cA6wsg8bbs8kqc9yU2xfosZkzbbYlJq5Ey3bok84R5NZSQS-YI2IDcjMgx-v2A__LEVtmkktf_yRHiykqxvdDvQmXNThTxOQkeh8_9weFqkmtIaFJHtlFujmRkE1-X790Na92qayG3QKvFZs9Vg2D2EeVq7CLfAeHXS77YFbLNSTme7WW97PMhgBBTBHaQfQUNBVkNYuHZiZ7c",
                      "proofPurpose": "assertionMethod",
                      "type": "JsonWebSignature2020",
                      "verificationMethod": "did:web:compliance.lab.gaia-x.eu"
                  },
                  "type": [
                      "VerifiableCredential"
                  ]
              },
              {
                  "@context": [
                      "https://www.w3.org/2018/credentials/v1",
                      "https://www.w3.org/2018/credentials/examples/v1"
                  ],
                  "credentialSubject": {
                      "@context": {
                          "example": "https://github.com/GAIA-X4PLC-AAD/ontology-management/tree/main/examples/example_ontology/",
                          "openlabel": "https://openlabel.asam.net/V1-0-0/ontologies/",
                          "sh": "http://www.w3.org/ns/shacl#",
                          "skos": "http://www.w3.org/2004/02/skos/core#",
                          "xsd": "http://www.w3.org/2001/XMLSchema#"
                      },
                      "@id": "asdfasd",
                      "@type": "example:ExampleServiceOffering",
                      "example:belongs_to": {
                          "@id": "fdgfdd"
                      },
                      "example:hasJunctionIntersection": {
                          "@id": "dfgdg"
                      },
                      "example:property1": {
                          "@type": "xsd:string",
                          "@value": "dfghjt"
                      },
                      "example:property2": {
                          "@type": "xsd:string",
                          "@value": "ththththt"
                      }
                  },
                  "expirationDate": "2022-06-18T12:30:00Z",
                  "issuanceDate": "2022-01-01T12:30:00Z",
                  "issuer": "did:web:compliance.lab.gaia-x.eu",
                  "proof": {
                      "created": "2022-01-01T12:30:00Z",
                      "jws": "eyJiNjQiOmZhbHNlLCJjcml0IjpbImI2NCJdLCJhbGciOiJQUzI1NiJ9..ZEsy_Y_psm67lu7Ie48eLqNTCIfCl0L1oYe4XPYCk36C4mRDhyuLI8oQ6EyZ8_AxhYhlV25lLIMZK6wpj3KjHOH3n3mvGW4qLalslly2n_0hal3CN8JZ2wlopCoOGy2-q-Jh6n1u9-glhQSDBzCngSlV0iFvP5IU-mcUDlvoQ52MzqhWtHGxEe40MS0N_GBwFiWmvCCj1PmGZeEyJoRolsIVKKIkvhQ3mEvrjx56-eoebZByUN9JPBzopkGEd7Y5hq4ppDpZEXAPS8a2zhYi0eNHIojkunxgcgD8PZ90IGky0Z8QKgokzXsvkajd8iqT2cki-1kn0I2F_L8B-3bfTqT4W1MuG-2LWDgG7nyOj4wbhkzJ2zL-6obsRMTAxfa4JjcTOoy29_S2GAuwav08kzJpv57ZV5XilCntJh0pR0du4fJxEvgm-0olB6kgFhgugzbnvbHzhYBk8212Mr0f6vuTt3b8cZNj9NgwryGAojqe-4q8dXhYG8GsG_Ou54DcLRbo_17T6QF7DeA6Ud35EyavsmsC7eAeQcjoM5hKYFYFsMfTS1FvZRjUNKnwBcnMKT9BxlXV1uxF8a3AsryEsnhdUwoblNm-bAGk6e-FRRo72nU3psWzbiBxcR-IvvtC6OfXfwvnSLOZ9eOQtCctp1k6RN363jpup3dS4cxscj0",
                      "proofPurpose": "assertionMethod",
                      "type": "JsonWebSignature2020",
                      "verificationMethod": "did:web:compliance.lab.gaia-x.eu"
                  },
                  "type": [
                      "VerifiableCredential"
                  ]
              }
              ]
      responses:
        '200': # status code
          description: The created Self Description.
          content:
            application/json:
              schema:
                type: object
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object

  /id-documents:
    get:
      summary: Returns all UUIDs (max. 500) of the objects that have been saved in the did_storage of the application.
      description: Returns all saved UUIDs (max. 500) in the did_storage as a list. 
      parameters:
        - in: query
          name: uuid
          required: false
          schema:
            type: str
          description: Provide a UUID to get the content of the regarding object.
      responses:
        '200': # status code
          description: The requested documents.
          content:
            application/json:
              schema:
                type: object
        '500': # status code
          description: In case an error occurred.
          content:
            application/json:
              schema:
                type: object